% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-analysis.R
\name{format.summary.trendanalysis}
\alias{format.summary.trendanalysis}
\title{Format a trend analysis summary for pretty-printing}
\usage{
\method{format}{summary.trendanalysis}(x, ...)
}
\arguments{
\item{x}{any \R object (conceptually); typically numeric.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Format method for summary.trendanalysis class
}
