% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCAManalysis.R
\docType{package}
\name{GCAManalysis}
\alias{GCAManalysis}
\alias{GCAManalysis-package}
\title{GCAManalysis: Analysis tools and algorithms for GCAM data}
\description{
The GCAManalysis package provides a variety of tools for analyzing GCAM data.
}
\section{Trend Analysis}{


The \code{\link{trend_analysis}} function detects and categorizes trends in
the data.  The "trend-analysis" vignette gives an extended example of using
this function.
}

