% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-analysis.R
\name{is.trendanalysis}
\alias{is.trendanalysis}
\title{Is an object a tendanalysis object}
\usage{
is.trendanalysis(object)
}
\arguments{
\item{object}{any \R object.}
}
\description{
Is method for trendanalysis class
}
