% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-analysis.R
\name{plot.trendanalysis}
\alias{plot.trendanalysis}
\title{Plot a trend analysis}
\usage{
\method{plot}{trendanalysis}(x, y, ...)
}
\arguments{
\item{x}{The trendanalysis object to plot}

\item{y}{ignored}

\item{...}{ignored}
}
\description{
Plot method for the trendanalysis class. Note that we use ggplot to actually
render the plot, not base graphics.
}
