% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-analysis.R
\name{print.summary.trendanalysis}
\alias{print.summary.trendanalysis}
\title{Print a trend analysis summary}
\usage{
\method{print}{summary.trendanalysis}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for summary.trendanalysis class
}
