% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-analysis.R
\name{summary.trendanalysis}
\alias{summary.trendanalysis}
\title{Summarize a trend analysis}
\usage{
\method{summary}{trendanalysis}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary method for trendanalysis class.
}
\details{
The summary includes the number of trend categories and the table of category
identifications.
}
